Attempt to fix STL_verify error
Tried to change it to a unique_ptr inorder to deal with broken references, but that didnt work. will have to change implementation to a vector which still gives me o(1) access. Will update past this week given the chance as this project isnt done but I dont want to extend this past 2 weeks.