Fix Deref end itter error
After many hours of struggling ive decided it cant throw this error if there is no itter to dereference.

Ive switched to a shared ptr to not have to use std::move()
Ive replaced the rendering code loops with ranged based for loops for better reading and no derefrencing
Ive replaced the findObject method to use at rather than find(). While this does cause crashes if the object is not found it overall plays better than the itterators